build: ensure build assets are versioned
This update will ensure that the required build assets as built in the
remote will be committed back to the project negating the need for a
developer to commit after a build.

On branch dsande/fallback
Your branch is up to date with 'origin/dsande/fallback'.
Changes to be committed:
modified:   package.json